modules/api-locations-producer/requirements.txt
grpcio == 1.46.0
grpcio-tools == 1.46.0
kafka-python == 2.0.2
